bug fix: undo of a multi-angle rotate does not undo, it just screws up the rotation completely.
The bug was that rotate was negating the eulers to "undo" the rotate.
But this is of coarse wrong - the undo has to be the negative angles
in the opposite order.

The new code uses a utility to calculate the inverse transform and
calculate new Eulers that represent it.

The name V3AntiEuler is a good one - anyone who is posing models
should be anti-Euler!