bug fix: part migration bakes in origin movement.
To test: put part 3855.dat (the old 1x4x3 glass) into a train window
6556.dat (the 1x4x3 train window).  If you're as old as me you
remember this from  Space lego sets.  Give them both the same
origin, and they'll line up perfectly.  Save it!

When you reload, let Bricksmith update the part - in old versions
the glass will be floating half way above the window, due to the
origin change (the old glass had the origin of the window frame at
the top, standard for bricks; the new glass has it's origin around
the hing point).

In the new code the glass stays in the window, but rotating with the
XYZ keys will show the new part has a center origin.